Refactor: Extract new workspace sftp-test-common
