+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "openssh-portable"]
-	path = openssh-portable
+	path = sftp-test-common/openssh-portable
 	url = https://github.com/openssh/openssh-portable
