+++ Cargo.toml
@@ -14,6 +14,11 @@ categories = ["asynchronous", "network-programming", "api-bindings"]
 
 exclude = ["openssh-portable"]
 
+[workspace]
+members = [
+    "sftp-test-common",
+]
+
 [features]
 # This feature is for internal testing only!!!
 ci-tests = []
@@ -41,9 +46,8 @@ pin-project = { version = "1.0.10" }
 tokio-pipe = "0.2.10"
 
 [dev-dependencies]
-tokio = { version = "1.11.0", features = ["rt", "macros", "process"] }
-once_cell = "1.8.0"
+tokio = { version = "1.11.0", features = ["rt", "macros"] }
 tempfile = "3.1.0"
 pretty_assertions = "1.1.0"
-libc = "0.2.112"
 tokio-pipe = "0.2.10"
+sftp-test-common = { path = "sftp-test-common" }
