+++ sftp-test-common/Cargo.toml
@@ -1,0 +1,15 @@
+# This package is meant for testing only
+# and not intended to be published to crates.io
+[package]
+name = "sftp-test-common"
+version = "0.1.0"
+edition = "2018"
+
+authors = ["Jiahao XU <Jiahao_XU@outlook.com>"]
+license = "MIT"
+
+[dependencies]
+tokio = { version = "1.11.0", features = ["process"] }
+once_cell = "1.8.0"
+libc = "0.2.112"
+tokio-pipe = "0.2.10"
